backend/requirements/dev.txt
Django==3.1.6
django-environ==0.4.5
django-model-utils==4.1.1
djangorestframework==3.12.2
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
cryptography==35.0.0
whitenoise
psycopg2
django-bootstrap-v5
pillow
openpyxl
gunicorn==20.1.0
